Tender_API/telegram_bot/requirements.txt
python-telegram-bot==20.8
httpx
python-dotenv
nest_asyncio

